提前了UI Command Queue的初始化时间，防止没有初始化的情况下调用了ui_cmd导致出错 修正了UTK命令数据长度用byte表示，不能达到最大的UIM卡字符范围，改成word类型
git-svn-id: svn://127.0.0.1@685 3f3bbce5-bc85-8142-9eb1-77d090cfebd4

+++ main/platform/mmi/core/ui.h
@@ -90,7 +90,7 @@ typedef struct
 typedef struct
 {
     ui_hdr_type                hdr;                       /* header */
-    byte                       num_bytes;                 /* length */
+    uint16                     num_bytes;                 /* length */
     byte                       cmd_data[UIM_MAX_CHARS];   /* data */
 } ui_proactive_uim_cmd_type;
 #endif
